requirements.txt
streamlit>=1.30.0
openai>=1.12.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
